refactor: move the ossfragment to the settings module
The OssFragment is the easier place to start with moving settings out, so move that over to the new
Settings module
