+++ app-settings/src/main/res/layout/fragment_oss.xml
