+++ app/build.gradle
@@ -61,6 +61,7 @@ dependencies {
     implementation project(path: ':account')
     implementation project(path: ':app-discover')
     implementation project(path: ':app-profile')
+    implementation project(path: ':app-settings')
     implementation project(path: ':app-timeline')
     implementation project(path: ':core')
     implementation project(path: ':database')
@@ -74,7 +75,7 @@ dependencies {
     implementation 'androidx.browser:browser:1.0.0'
     implementation "androidx.constraintlayout:constraintlayout:$constraint_layout_version"
     implementation "androidx.core:core-ktx:$corektx_version"
-    implementation 'androidx.fragment:fragment-ktx:1.2.0-beta01'
+    implementation "androidx.fragment:fragment-ktx:$fragmentctx_version"
     implementation "androidx.lifecycle:lifecycle-extensions:$lifecycle_version"
     implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:$lifecycle_version"
     implementation "androidx.navigation:navigation-fragment-ktx:$nav_version"
@@ -95,7 +96,6 @@ dependencies {
     implementation "com.google.dagger:dagger-android:$dagger_version"
     implementation "com.google.dagger:dagger-android-support:$dagger_version"
     implementation "com.jakewharton.timber:timber:$timber_version"
-    implementation 'com.mikepenz:aboutlibraries:7.0.3'
     implementation "com.squareup.moshi:moshi:$moshi_version"
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:$coroutines_version"
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:$coroutines_version"
