+++ build.gradle
@@ -35,6 +35,7 @@ detekt {
             "app/src/main/java",
             "app-discover/src/main/java",
             "app-profile/src/main/java",
+            "app-settings/src/main/java",
             "app-timeline/src/main/java",
             "core/src/main/java",
             "database/src/main/java",
@@ -65,6 +66,7 @@ ext {
     corektx_version = "1.2.0-beta01"
     coroutines_version = "1.3.2"
     dagger_version = "2.24"
+    fragmentctx_version = "1.2.0-beta01"
     glide_version = "4.10.0"
     lifecycle_version = "2.2.0-beta01"
     lifecykle_version = "2.0.0"
